Set user.extension-release.strict=0 on the extension-release file
At some point we might revisit this, but in development keeping the
file name in sync with the internal name is a PITA. In particular, the
external name will usually include the kernel version, which means that the
internal name cannot be set in a configuration file. Maybe we can find some more
flexible mechanism in the future.