add kernel-install files
60-mkosi-initrd.install builds the initrd using mkosi
and puts it into a proper place.

Since we might want to run in parallel with dracut, the script checks
if dracut is present and in such case it creates an initrd
with a different suffix.

But in such case, we also need to also create a new BLS entry and that's
why we also have 95-mkosi-initrd-loaderentry.install. This ones copies
the entry from dracut based initrd and edits the initrd line.