Use $KERNEL_INSTALL_INITRD_GENERATOR and $KERNEL_INSTALL_STAGING_AREA
The first variable is used to check if we should do anything at all.
If it's not set to "mkosi-initrd", bow out immediately.

The second variable specifies a directory where to stash our initrd.
kernel-install will copy it into the appropriate location later on.
Since all files called "initrd*" are assumed to be initrds, let's not
generate the manifests in the same location.

Similarly, the output directory is unset, since mkosi warns if it is
set and --output is specified with a full path.

This way the plugin does just one thing — generates the initrd — and
doesn't have to worry about anything else.

kernel-install should take care of cleaning up the initrd on removal,
so let's skip that too.

Also, use the same number that 50-dracut.install uses. Let's say that
50 is the start level for initrd generators.