mkosi.finalize: when installing modules, copy them from the host fs
The obvious advantage is that we don't need to download the kernel rpm.
A subtler advantage is that we would often fail when downloading the
modules for a kernel which has since been superseded by a new version
and is not present on the package mirrors.

The obvious disadvantage is that the files on disk might not be in
pristine condition. A less obvious disadvantage is that we copy in
all the files that are there, including modules not from kernel-core,
so we end up with a huuuge initrd.