mkosi.finalize: also skip files in /doc/
The CI fails with:
FileNotFoundError: [Errno 2] No such file or directory: '/usr/share/doc/kernel-keys/5.19.0-65.fc37.x86_64'

It'd be nicest to exclude %doc files from the rpm list, but it seems
that rpm offers no mechanism to list file "class". The best I could
find is 'rpm -Vv' which prints the class for missing files. But maybe
excluding /doc/ is good enough as a work-around.